Assorted warning fixes, small mixer improvement when both DTK and HLE audio are used
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1100 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/ChunkFile.h
@@ -92,7 +92,7 @@ public:
 	// Store strings.
 	void Do(std::string &x) 
 	{
-		int stringLen = x.length() + 1;
+		int stringLen = (int)x.length() + 1;
 		Do(stringLen);
 
 		switch (mode)
